+++ backend/app/models/assignment.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import enum
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Optional
 
 from sqlalchemy import Date, DateTime, Enum, ForeignKey, Integer, Numeric, String, func
@@ -19,14 +20,22 @@ class Assignment(Base):
     __tablename__ = "assignments"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    employee_id: Mapped[int] = mapped_column(Integer, ForeignKey("employees.id"), nullable=False, index=True)
-    project_id: Mapped[int] = mapped_column(Integer, ForeignKey("projects.id"), nullable=False, index=True)
+    employee_id: Mapped[int] = mapped_column(
+        Integer, ForeignKey("employees.id"), nullable=False, index=True
+    )
+    project_id: Mapped[int] = mapped_column(
+        Integer, ForeignKey("projects.id"), nullable=False, index=True
+    )
     start_date: Mapped[date] = mapped_column(Date, nullable=False)
     end_date: Mapped[date] = mapped_column(Date, nullable=False)
-    allocation_type: Mapped[AllocationType] = mapped_column(Enum(AllocationType), nullable=False)
-    allocation_value: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
+    allocation_type: Mapped[AllocationType] = mapped_column(
+        Enum(AllocationType), nullable=False
+    )
+    allocation_value: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False)
     note: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True), server_default=func.now()
+    )
     updated_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), onupdate=func.now()
     )
